+++ src/widgets/settingspages/GeneralPage.cpp
@@ -505,8 +505,8 @@ void GeneralPage::initLayout(GeneralPageView &layout)
                      "message to help better tell them apart.")
         ->addTo(layout);
 
-    SettingWidget::checkbox("Reduce opacity of past messages",
-                            s.fadeRecentMessages)
+    SettingWidget::checkbox("Reduce opacity of message history",
+                            s.fadeMessageHistory)
         ->setTooltip(
             "Reduce opacity of messages that were posted before Chatterino "
             "was started or while re-connection.")